requirements.txt
APScheduler==3.6.3
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.11.2
drf-oidc-auth==0.10.0
drf-yasg==1.17.1
Jinja2==3.0.0a1
jsonschema==3.2.0
mock==4.0.2
model-bakery==1.2.0
protobuf==4.0.0rc2
psycopg2-binary==2.8.6
PyYAML==5.3.1
rpyc==4.1.5
schema==0.7.3
structlog==20.1.0
pytest==6.1.1
pytest-cov==2.10.1
pytest-django==4.0.0
pytz==2020.1
amqpstorm==2.8.2
uuid==1.30
sqlalchemy==1.3.20